feat: create a map with port and corresponding servers 	add upload_allowed and upload_location docs: write the order of priority while reading a request fix: errors

+++ includes/parser/ServerConfig.hpp
@@ -34,7 +34,7 @@ public:
 	void	setIndex( const std::string& index );
 	void	setErrorPage( int error_nb, const std::string& error_page );
 	void	setAutoindex( AutoindexType autoindex_ );
-	void	setClientMaxBodySize( unsigned long client_max_body_size );
+	void	setClientMaxBodySize( long long client_max_body_size );
 	void	setReturn( int return_code, const std::string& return_page );
     void	setHasReturn();
 
@@ -49,7 +49,7 @@ public:
 	const std::vector<std::string>&	    getIndex() const;
 	const std::map<int, std::string>&	getErrorPage() const;
 	AutoindexType		 		            getAutoindex() const;
-	unsigned long			getClientMaxBodySize() const;
+	long long			getClientMaxBodySize() const;
 	const std::map<int, std::string>&	getReturn() const;
     bool					getHasReturn() const;
 
@@ -65,7 +65,7 @@ private:
 	std::vector<std::string>	index_;
 	std::map<int, std::string>	error_page_;
 	AutoindexType				autoindex_;
-	unsigned long			client_max_body_size_;
+	long long			client_max_body_size_;
 	std::map<int, std::string>	return_code_;
 };
